refactor(client): :art: refactor layout
